# web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "F1 Data Platform API",
    "summary": "Resultados, vueltas, neumáticos, telemetría y estadísticas de la F1 (1950-hoy).",
    "description": "API de solo lectura sobre el modelo dimensional del proyecto. Fuentes: F1DB, FastF1, OpenF1, formula1db.com y Ergast (solo validación), conciliadas con documentos de la FIA. Los tiempos van en milisegundos y los identificadores son los de F1DB (p. ej. `lewis-hamilton`, `ferrari`).\n\nCódigo MIT. Los datos conservan la licencia de su fuente: F1DB (CC BY 4.0), Ergast (CC BY-NC-SA 3.0), OpenF1 (CC BY-NC-SA 4.0: lo que lleva `source = openf1`), el cronometraje de la F1 obtenido con FastF1 (© Formula One, uso académico no comercial) y formula1db.com (con permiso de su autor); ver [licencias de los datos](https://github.com/AntonioSchez32/f1-data-platform#licencia-y-atribuciones). Cita «F1 Data Platform» y la fuente original. Proyecto académico no oficial, sin relación con la Fórmula 1 ni con la FIA.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/drivers": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "summary": "Buscar pilotos",
        "operationId": "list_drivers_drivers_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "Parte del nombre",
              "title": "Search"
            },
            "description": "Parte del nombre"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Solo pilotos de esta temporada",
              "title": "Season"
            },
            "description": "Solo pilotos de esta temporada"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverSummary"
                  },
                  "title": "Response List Drivers Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/constructors": {
      "get": {
        "tags": [
          "Constructores"
        ],
        "summary": "Buscar constructores",
        "operationId": "list_constructors_constructors_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "Parte del nombre",
              "title": "Search"
            },
            "description": "Parte del nombre"
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Solo constructores de esta temporada",
              "title": "Season"
            },
            "description": "Solo constructores de esta temporada"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConstructorSummary"
                  },
                  "title": "Response List Constructors Constructors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/records/drivers": {
      "get": {
        "tags": [
          "Récords"
        ],
        "summary": "Ranking histórico de pilotos",
        "operationId": "driver_records_records_drivers_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "championships",
                "wins",
                "podiums",
                "pole_positions",
                "fastest_laps",
                "race_starts",
                "points",
                "grand_slams"
              ],
              "type": "string",
              "default": "wins",
              "title": "Metric"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordEntry"
                  },
                  "title": "Response Driver Records Records Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/records/constructors": {
      "get": {
        "tags": [
          "Récords"
        ],
        "summary": "Ranking histórico de constructores",
        "operationId": "constructor_records_records_constructors_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "championships",
                "wins",
                "podiums",
                "pole_positions",
                "fastest_laps",
                "race_entries",
                "one_two_finishes",
                "points",
                "points_historical"
              ],
              "type": "string",
              "default": "wins",
              "title": "Metric"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordEntry"
                  },
                  "title": "Response Constructor Records Records Constructors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rankings/drivers": {
      "get": {
        "tags": [
          "Récords"
        ],
        "summary": "Pilotos más laureados",
        "operationId": "driver_rankings_rankings_drivers_get",
        "parameters": [
          {
            "name": "season_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Desde esta temporada",
              "title": "Season From"
            },
            "description": "Desde esta temporada"
          },
          {
            "name": "season_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hasta esta temporada",
              "title": "Season To"
            },
            "description": "Hasta esta temporada"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "wins",
                "championships",
                "podiums",
                "pole_positions",
                "fastest_laps",
                "points",
                "entries"
              ],
              "type": "string",
              "default": "wins",
              "title": "Order By"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DriverRanking"
                  },
                  "title": "Response Driver Rankings Rankings Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rankings/constructors": {
      "get": {
        "tags": [
          "Récords"
        ],
        "summary": "Constructores más laureados",
        "operationId": "constructor_rankings_rankings_constructors_get",
        "parameters": [
          {
            "name": "season_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Desde esta temporada",
              "title": "Season From"
            },
            "description": "Desde esta temporada"
          },
          {
            "name": "season_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hasta esta temporada",
              "title": "Season To"
            },
            "description": "Hasta esta temporada"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "wins",
                "championships",
                "podiums",
                "pole_positions",
                "fastest_laps",
                "points",
                "points_historical",
                "entries"
              ],
              "type": "string",
              "default": "wins",
              "title": "Order By"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConstructorRanking"
                  },
                  "title": "Response Constructor Rankings Rankings Constructors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConstructorRanking": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "country_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Alpha2"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "podiums": {
            "type": "integer",
            "title": "Podiums"
          },
          "pole_positions": {
            "type": "integer",
            "title": "Pole Positions"
          },
          "fastest_laps": {
            "type": "integer",
            "title": "Fastest Laps"
          },
          "championships": {
            "type": "integer",
            "title": "Championships"
          },
          "points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points",
            "description": "Carrera y sprint de las temporadas desde 1958 del rango (F1DB); nulo si el rango no incluye ninguna en que corriera"
          },
          "points_historical": {
            "type": "number",
            "title": "Points Historical",
            "description": "Carrera y sprint de todo el rango"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "id",
          "name",
          "country",
          "country_alpha2",
          "entries",
          "wins",
          "podiums",
          "pole_positions",
          "fastest_laps",
          "championships",
          "points",
          "points_historical"
        ],
        "title": "ConstructorRanking"
      },
...
      "ConstructorSummary": {
        "properties": {
          "constructor_id": {
            "type": "string",
            "title": "Constructor Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Alpha2"
          },
          "first_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Season"
          },
          "last_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Season"
          },
          "race_entries": {
            "type": "integer",
            "title": "Race Entries"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "championships": {
            "type": "integer",
            "title": "Championships"
          }
        },
        "type": "object",
        "required": [
          "constructor_id",
          "name",
          "country_alpha2",
          "first_season",
          "last_season",
          "race_entries",
          "wins",
          "championships"
        ],
        "title": "ConstructorSummary"
      },
...
      "DriverRanking": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "country_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Alpha2"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "podiums": {
            "type": "integer",
            "title": "Podiums"
          },
          "pole_positions": {
            "type": "integer",
            "title": "Pole Positions"
          },
          "fastest_laps": {
            "type": "integer",
            "title": "Fastest Laps"
          },
          "championships": {
            "type": "integer",
            "title": "Championships"
          },
          "points": {
            "type": "number",
            "title": "Points"
          },
          "starts": {
            "type": "integer",
            "title": "Starts"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "id",
          "name",
          "country",
          "country_alpha2",
          "entries",
          "wins",
          "podiums",
          "pole_positions",
          "fastest_laps",
          "championships",
          "points",
          "starts"
        ],
        "title": "DriverRanking"
      },
...
      "DriverSummary": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nationality_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality Alpha2"
          },
          "first_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Season"
          },
          "last_season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Season"
          },
          "race_starts": {
            "type": "integer",
            "title": "Race Starts"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "podiums": {
            "type": "integer",
            "title": "Podiums"
          },
          "championships": {
            "type": "integer",
            "title": "Championships"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "name",
          "nationality_alpha2",
          "first_season",
          "last_season",
          "race_starts",
          "wins",
          "podiums",
          "championships"
        ],
        "title": "DriverSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecordEntry": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country_alpha2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Alpha2"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "id",
          "name",
          "country_alpha2",
          "value"
        ],
        "title": "RecordEntry"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}